+++ src/layouts/adminLayout/AdminLayout.tsx
@@ -1,62 +1,27 @@
 import React from "react";
 import { styled } from "styled-components";
 import Footer from "../footer/Footer";
-import { Link, NavLink, useLocation } from "react-router-dom";
+import { Link } from "react-router-dom";
 
 import Logo from "../../assets/logo.svg";
+import { AdminNav } from "./AdminNav";
 
 type WrapperProps = {
   children: React.ReactNode;
 };
 
 const AdminLayout = ({ children }: WrapperProps) => {
-  const deactiveStyle = {
-    fontFamily: "var(--font--Medium)",
-    color: "var(--gray500-color)",
-  };
-  const activeStyle = {
-    fontFamily: "var(--font--Bold)",
-    color: "var(--Black-color)",
-    borderBottom: "2px solid var(--main-color)",
-  };
-
-  const location = useLocation();
-  const isActiveManagementPath = ["/admin/barinfo", "/admin/management"].includes(location.pathname);
-  console.log(isActiveManagementPath);
   return (
     <LayoutWrap>
       <FixHeader>
-        <H1>
+        <h1>
           <Link to="/admin/barinfo">
             <img src={Logo} alt="Zanit 관리자 페이지" />
           </Link>
-        </H1>
+        </h1>
         <strong>관리자 페이지</strong>
       </FixHeader>
-      <Navbar>
-        <ul>
-          <li>
-            <NavLink
-              to="/admin/barinfo"
-              style={() => {
-                return isActiveManagementPath ? activeStyle : deactiveStyle;
-              }}
-            >
-              정보 관리
-            </NavLink>
-          </li>
-          <li>
-            <NavLink
-              to="/admin/manageorder"
-              style={({ isActive }) => {
-                return isActive ? activeStyle : deactiveStyle;
-              }}
-            >
-              주문 관리
-            </NavLink>
-          </li>
-        </ul>
-      </Navbar>
+      <AdminNav />
       <ScrollMain>
         <main>{children}</main>
         <Footer />
@@ -73,6 +38,10 @@ const FixHeader = styled.header`
   top: 0;
   background-color: var(--white-color);
 
+  h1 {
+    padding: 13px;
+  }
+
   strong {
     position: absolute;
     left: 50%;
@@ -80,10 +49,6 @@ const FixHeader = styled.header`
     transform: translate(-50%, -50%);
     font-family: var(--font--Bold);
   }
-`;
-
-const H1 = styled.h1`
-  padding: 13px;
 `;
 
 const LayoutWrap = styled.div`
@@ -114,21 +79,5 @@ const ScrollMain = styled.div`
   &::-webkit-scrollbar-thumb {
     background-color: #b7b7b7;
     border-radius: 5px;
-  }
-`;
-
-const Navbar = styled.nav`
-  padding-bottom: 20px;
-  ul {
-    display: flex;
-    flex-direction: row;
-    gap: 20px;
-    padding: 0 15px;
-
-    a {
-      display: inline-block;
-      padding: 15px 0 8px;
-      font-size: 15px;
-    }
   }
 `;
