+++ src/libs/apis/user.ts
@@ -1,7 +1,6 @@
 // 유저 로그인, 회원가입 API
 
 import { defaultInstance, formDataInstance } from "./axios";
-import axios from "axios";
 
 interface signUpUser {
   userPhone: string;
@@ -35,6 +34,6 @@ export const signInAPI = async (formData: FormData): Promise<ApiResponse> => {
     return { status: res.status, data: res.data };
   } catch (e) {
     console.error(e);
-    throw e;
+    return { status: 500, data: "api 호출 실패" };
   }
 };
